[tests][unit] Updated dee tests, disabled until NOELLE is updated
