--- tests/unit/dee_sequence_insert/optflags
@@ -1,1 +1,0 @@
--passes=memoir-dee
