--- tests/unit/dee_sequence_readwrite/optflags
@@ -1,1 +1,0 @@
--passes=memoir-dee
